make generic scan error base widget public
